Finished out export and error_mols handling for failed optimizations
Implementation not quite as condensed as I'd like, but this is getting a
bit tricky. One alternative is to ditch `error_mols` entirely and take
the convention we've used so far that failed molecules get everything
but SDF output, so their results don't make it downstream but error
output still introspectable.