[WIP] Detect proton transfer and add error handling logic to QM export (#65)
* add test for catching connectivity rearrangement

* add example file for connectivity rearrangement

* add method to detect connectivity rearrangement

* first draft of changes for error-tolerant export

* add error for connectivity rearrangement

* fix test

* fix connectivity check

* do connectivity check in molecule processing

* fix staticmethod signature

* Working on error handling; realized we probably don't want error_mols dir

If we hit a problem with an optimization or with final molecule
processing, we don't want an SDF file at all. Our current approach
yields every other file except the SDF, and I'm in favor of preserving
this.

What's more, I *think* some of the analysis machinery does recursive
dives...but if this doesn't hit SDF files, should be okay.

Think I may have talked myself into doing `error_mols` after all
for consistency with previous pipeline components.

* Finished out export and error_mols handling for failed optimizations

Implementation not quite as condensed as I'd like, but this is getting a
bit tricky. One alternative is to ditch `error_mols` entirely and take
the convention we've used so far that failed molecules get everything
but SDF output, so their results don't make it downstream but error
output still introspectable.

* Test coverage for proton transfer failure paths

Co-authored-by: David Dotson <dotsdl@gmail.com>