Added try-except for rdMolAlign call, NaN injection on failure with message
This is in response to an issue Xavier Lucas ran into in which a failure
to calculate the RMSD by RDKit causes the whole analysis to fail.
This change makes the analysis tolerant of failures at this point, while
giving an informative message indicating on which conformer(s) the failure occurred.